third_party/blink/web_tests/platform/mac-mac12/external/wpt/css/css-scrollbars/scrollbar-color-002-expected.txt

This is a testharness.js-based test.
[FAIL] scrollbar-color non-auto on the root overrides ::-webkit-scrollbar
  assert_less_than: viewport has a scrollbar expected a number less than 1200 but got 1200
Harness: the test ran to completion.
